+++ practices/RefactoringToDelegates/Tests/ConvertingInchesToMilllimeters.cs
@@ -8,9 +8,9 @@ namespace Tests
 		[TestMethod]
 		public void Example1Inch()
 		{
-			const float inches = 1;
+			const double inches = 1;
 
-			float mm = _converter.ConvertInchesToMillimeters(inches);
+			double mm = _converter.ConvertInchesToMillimeters(inches);
 
 			Assert.AreEqual(25.4, mm, .01); // the third argument is a delta - how close does the float have to be. Comparing floats for equality is bad. See http://docs.oracle.com/cd/E19957-01/806-3568/ncg_goldberg.html
 
@@ -19,7 +19,7 @@ namespace Tests
 		[TestMethod]
 		public void Example1point5Inches()
 		{
-			const float inches = 1.5f;
+			const double inches = 1.5;
 
 			var mm = _converter.ConvertInchesToMillimeters(inches);
 
@@ -30,7 +30,7 @@ namespace Tests
 		[TestMethod]
 		public void BigInches()
 		{
-			const float inches = 128.385f;
+			const double inches = 128.385f;
 
 			var mm = _converter.ConvertInchesToMillimeters(inches);
